requirements.txt
# --- core web framework & config --------------------------------------
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.11.5

# --- LINE & Google stacks you use -------------------------------------
line-bot-sdk==3.9.0
google-genai==1.19.0
google-api-python-client==2.121.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
# gspread==5.12.4            # REMOVED - replaced by database logging
# oauth2client==4.1.3        # REMOVED - replaced by database logging

# --- other top-level imports in the code ------------------------------
dnspython==2.6.1           # MX-record check
beautifulsoup4==4.12.3     # extract Gemini grounding references
# pydub==0.25.1            # REMOVED - not used in codebase

# --- database dependencies for Neon DB logging ------------------------
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg>=0.29.0

# ----------------------------------------------------------------------
# Everything below is pulled in automatically by the libraries above
# OR is not referenced anywhere in the repo.  Leave them commented.
# ----------------------------------------------------------------------
#aenum==3.1.16
#aiohttp==3.9.3
#aiosignal==1.3.2
#annotated-types==0.7.0
#anyio==4.9.0
#attrs==25.3.0
#cachetools==5.5.2
#certifi==2025.4.26
#charset-normalizer==3.4.2
#click==8.2.1
#colorama==0.4.6
#Deprecated==1.2.18
#frozenlist==1.6.2
#future==1.0.0
#google-api-core==2.25.0
#googleapis-common-protos==1.70.0
#h11==0.16.0
#httpcore==1.0.9
httplib2==0.22.0
#httptools==0.6.4
#httpx==0.28.1
#idna==3.10
#multidict==6.4.4
#oauthlib==3.2.2
#propcache==0.3.1
#proto-plus==1.26.1
#protobuf==6.31.1
#pyasn1==0.6.1
#pyasn1_modules==0.4.2
#pydantic_core==2.33.2
#pyparsing==3.2.3
#python-dateutil==2.9.0.post0
#PyYAML==6.0.2
#requests==2.31.0
#requests-oauthlib==2.0.0
#rsa==4.9.1
#six==1.17.0
#sniffio==1.3.1
#soupsieve==2.7
#starlette==0.36.3
#typing-inspection==0.4.1
#typing_extensions==4.14.0
#uritemplate==4.2.0
#urllib3==2.4.0
#watchfiles==1.0.5
#Wave==0.0.2
#websockets==15.0.1
#wrapt==1.17.2
#yarl==1.20.0
